# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/todo":{"post":{"operationId":"TodoController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoDto"}}}},"responses":{"201":{"description":""}},"tags":["todo"]},"get":{"operationId":"TodoController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"withDeleted","required":false,"in":"query","schema":{"type":"boolean"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"orderKey","required":false,"in":"query","schema":{"type":"string"}},{"name":"orderValue","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["todo"]}},"/todo/{id}":{"get":{"operationId":"TodoController_findOne","parameters":[],"responses":{"200":{"description":""}},"tags":["todo"]},"patch":{"operationId":"TodoController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoDto"}}}},"responses":{"200":{"description":""}},"tags":["todo"]},"delete":{"operationId":"TodoController_remove","parameters":[],"responses":{"200":{"description":""}},"tags":["todo"]}},"/auth/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}}},"info":{"title":"todo example","description":"The todo API description","version":"1.0","contact":{}},"tags":[{"name":"todo","description":""}],"servers":[],"components":{"securitySchemes":{"ApiKeyAuth":{"scheme":"bearer","bearerFormat":"JWT","type":"apiKey","in":"header","name":"Authorization"}},"schemas":{"CreateTodoDto":{"type":"object","properties":{"task":{"type":"string","example":"task"}},"required":["task"]},"UpdateTodoDto":{"type":"object","properties":{"task":{"type":"string","example":"task"}}},"SigninDto":{"type":"object","properties":{"email":{"type":"string","example":"abdo@gmail.com"},"password":{"type":"string","example":"12345678"}},"required":["email","password"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"abdo"},"email":{"type":"string","example":"abdo@gmail.com"},"password":{"type":"string","example":"12345678"}},"required":["name","email","password"]}}},"security":[{"ApiKeyAuth":[]}]}